automation/requirements.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Frontend
streamlit==1.30.0
plotly==5.18.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Document Generation
reportlab==4.0.8
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
python-multipart==0.0.6
